refactor: split files by purpose
