+++ indexer.py
@@ -7,8 +7,8 @@ from decimal import Decimal, getcontext
 from sqlalchemy.exc import IntegrityError
 from web3.constants import ADDRESS_ZERO
 
-import database
 import settings
+from models import FundraisePositions, IndexerState, Session, StakingPositions, StakingPositionsMeta
 from utils import time_delta_apy
 
 YEAR = Decimal(timedelta(days=365).total_seconds())
@@ -29,7 +29,7 @@ class Indexer:
         if self.block_last_updated == block:
             return
 
-        meta = database.StakingPositionsMeta.get(session)
+        meta = StakingPositionsMeta.get(session)
         if meta.usdc_last_updated == datetime.min:
             self.block_last_updated = block
             return
@@ -39,7 +39,7 @@ class Indexer:
         if position_timedelta == 0:
             return
 
-        position = database.StakingPositions.get_for_factor(session)
+        position = StakingPositions.get_for_factor(session)
         usdc, factor = position.get_balance_data(block)
 
         # TODO make compounding?
@@ -56,11 +56,11 @@ class Indexer:
     class Transfer:
         def new(self, session, indx, block, args):
             if args["to"] == ADDRESS_ZERO:
-                database.StakingPositions.delete(session, args["tokenId"])
+                StakingPositions.delete(session, args["tokenId"])
                 return
             elif args["from"] != ADDRESS_ZERO:
                 # from and to both are not zero, this is an actual transfer
-                database.StakingPositions.update(
+                StakingPositions.update(
                     session,
                     args["tokenId"],
                     args["to"],
@@ -71,11 +71,11 @@ class Indexer:
             self.calc_factors(session, indx, block)
 
             if indx.balance_factor != Decimal(1):
-                database.StakingPositionsMeta.update(session, block, indx.balance_factor)
+                StakingPositionsMeta.update(session, block, indx.balance_factor)
                 indx.balance_factor = Decimal(1)
 
             # Insert will retrieve active information (usdc, sher, lockup)
-            database.StakingPositions.insert(
+            StakingPositions.insert(
                 session,
                 block,
                 args["tokenId"],
@@ -84,9 +84,9 @@ class Indexer:
 
     class Purchase:
         def new(self, session, indx, block, args):
-            position = database.FundraisePositions.get(session, args["buyer"])
+            position = FundraisePositions.get(session, args["buyer"])
             if position is None:
-                database.FundraisePositions.insert(
+                FundraisePositions.insert(
                     session,
                     block,
                     args["buyer"],
@@ -101,14 +101,14 @@ class Indexer:
 
     def start(self):
         # get last block indexed from database
-        with database.Session() as s:
-            start_block = s.query(database.IndexerState).first().last_block
+        with Session() as s:
+            start_block = s.query(IndexerState).first().last_block
 
         t = threading.currentThread()
         logging.debug("Thread started")
         while getattr(t, "do_run", True):
             try:
-                s = database.Session()
+                s = Session()
 
                 # Process 5 blocks each round
                 end_block = start_block + settings.INDEXER_BLOCKS_PER_CALL - 1
@@ -122,7 +122,7 @@ class Indexer:
                     continue
 
                 # If think update apy factor here? So we can already use in the events
-                indx = s.query(database.IndexerState).first()
+                indx = s.query(IndexerState).first()
 
                 self.index_events_time(s, indx, start_block, end_block)
                 self.calc_factors(s, indx, end_block)
