+++ views/__init__.py
@@ -1,0 +1,2 @@
+from .fundraise import *  # noqa
+from .staking import *  # noqa
